+++ Engine/Window.cpp
@@ -145,7 +145,7 @@ namespace Bat
 
 			if( msg.message == WM_QUIT )
 			{
-				PostQuitMessage( msg.wParam );
+				PostQuitMessage( (int)msg.wParam );
 				return false;
 			}
 		}
